Update patch set 1
Patch Set 1:

Thank you for this.

Given the context, I think https://gerrit.wikimedia.org/r/171024 is embarrassing and I hope it's never repeated. For a change to the biggest Wikimedia wiki, the lackluster commit message and seeming non-existent related public discussion (on wikitech-l or on Bugzilla or on the wiki itself or anywhere else of record...) about the implications of such a change should have been clear warning signs to not push the change to production. I'm boggling a bit.

Patch-set: 1
Reviewer: MZMcBride <531@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0