Add SecurePoll specific dblist
To allow for the separate manipulation of what is considered a 'global' election for SecurePoll purposes. 
Initially needed because of Bug: 72569 which requires us to get rid of labswiki for now but also good to have regardless because of wikis like loginwiki,zerowiki which should not be part of the main voting system. 
For now leaving decisions about chapter wikis etc for the board election committee later.

Bug: 73245
Change-Id: I831203bdb5db7df79032edaf447be41ecf55c287