Update patch set 3
Patch Set 3:

Actually getting it configured in SecurePoll is a little more work, but not much. 

* Add something like this into CommonSettings.php, in the block where other SecurePoll settings are set:

	$wgSecurePollCreateWikiGroups = array(
		'securepollglobal' => 'securepoll-dblist-securepollglobal'
	);

* Add an entry for the message key 'securepoll-dblist-securepollglobal' into the WikimediaMessages extension (i18n/wikimedia/en.json and i18n/wikimedia/qqq.json, probably).

If you want to do that in time for the Wednesday morning SWAT window I can SWAT it then, or if you tell me the name for the UI I can set it all up, probably still in time for that window.

Patch-set: 3
Label: Code-Review=0