Update patch set 2
Patch Set 2:

Temporarily... for two months?

I'm sorry you didn't see this, I had linked the change on Meta-Wiki and emailed wikitech-ambassador. Disabling the experiment doesn't preclude any future enabling; here we're just applying the original decision, that the experiment be temporary, for a meaningful value of "temporary".

And the original decision was applied because it's not clear who is currently responsible of renewing or revising it, probably nobody; in fact there wasn't even a bugzilla report about enabling the code, so it was impossible to properly discuss it.

Patch-set: 2
Reviewer: Nemo bis <105@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0