Send more update jobs to Elasticsearch
We rate limited the jobs for a while and it worked well.  At some point
a few weeks ago something changed and we could no longer keep up with
the rate limit.  Its not clear if that was caused by higher rate of change
or changes in how the rate limit works.  Either way, a higher rate limit
is in order.

This raises the rate limit to be a bit more than 5x what it was which seems
pretty safe.

Change-Id: Ic70bcbbda1c6ecb14bc62777dc22da46cf994c7e