Update patch set 3
Patch Set 3:

Right now the load average on the Elasticsearch cluster is super low.  I figured it was because we through so much more hardware at it.  I'm curious to see how this effects it.

Lots of updates can really slow things down.  If anything throws our plans for enwiki on Wednesday in jeopardy, it is this.

Patch-set: 3
Reviewer: Manybubbles <873@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0