Update patch set 2
Patch Set 2:

> Temporarily... for two months?

For only a small segment of users.  And no, we didn't figure out when to remove it (which is why I'm *not* reverting or saying it should never have been merged regardless of how long it sat here).

> And the original decision was applied because it's not clear who is currently responsible of renewing or revising it, probably nobody;

The original people that worked on it, all of whom are still here except Steven Walling.  Yes, we have less time for this now that the Growth team is retired, but that does not mean we're unresponsive.  I'm still on IRC, etc.

> in fact there wasn't even a bugzilla report about enabling the code, so it was impossible to properly discuss it.

How was this patch impossible to discuss?  It could be discussed here or in Bugzilla.

As for the issue of stuff being split between Bugzilla and Trello (there is stuff on Bugzilla related to this as well), if you notice I've been advocating for the migration to Phabricator to solve that problem for a long time.

Patch-set: 2