Make mdot webroot redirects agree with W0
Currently, W0 config changes from disabled to enabled may not
be reflected on the mdot webroot as fast as they are on the
Wikipedia Zero banners. This brings the Cache-Control headers in
line with those W0 banners. In addition to setting a 15 minute window
for the mdot webroot redirects, shared caches are now expressly
allowed to cache responses (already done implicitly it seems), and
an unnecessary must-revalidate directive is removed as it is implied
through use of s-maxage and its implicit proxy-revalidate.

A comment has been left in the file regarding cach variance for
the Accept-Language header, which will be coming in the following
patch, although that is not the subject of this very patch.

Change-Id: Icdd645c493b2118482c0ca24a662283b20454881