Update patch set 3
Patch Set 3:

As the security person, I'm not blocking this. The security issues have been addressed.

As a core team member, I think there are lots of issues with this that haven't been addressed, but someone who's been working on it would know better which ones are closed or not.

Patch-set: 3
Reviewer: CSteipp <204@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0