Rewrite robots.php
Robots.php was using long deprecated MW internal APIs. The change
rewrites it for more modern internal APIs, and (in the author's
opinion) makes the logic easier to follow by rearranging the logic and
renaming variables. The expected output is identical to the prior
version with the possible exception of HTTP header ordering.

Change-Id: Ibccd955e28a291c4050e6ccd5e81186674b7dfe0