Update patch set 5
Patch Set 5:

Eh, if we're rewriting robots.php, I'd kind of prefer that we centralize the rules by eliminating the /srv/mediawiki/robots.txt file. A few global (across Wikimedia wikis) unconditional rules may be needed, but as I recall, most of the rules from the text file are specific to a half-dozen projects (mostly the English Wikipedia) and we're currently applying them to hundreds of projects.

It may also make sense to write a few small tests to ensure that (for example) Zero and Beta still work as expected, given that the logic there seems a bit sensitive. This isn't strictly necessary, but might be nice to have.

Patch-set: 5
Reviewer: MZMcBride <531@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0