Allow enwiki bots to create flow boards
This enables two rights for the bots group on enwiki: editcontentmodel and
flow-create-board.  The result of this is that any bot can change a
non-existant page into a flow board by issuing an API request to create
a topic or edit the header of the page.

The initial user for this will be HostBot, an already approved bot, which
will be enabling specific sub pages for the Co-op[1] mentorship program.
We considered assigning the rights only to HostBot, but as the bots are
already a quite well behaved and predictable group it seems resonable to
assign the right to all bots.

[1]: https://en.wikipedia.org/wiki/Wikipedia_talk:Co-op

Bug: T76785
Change-Id: I4de5a9c330d2e56b0cd754030968e57fc8191434