beta: change monolog config to dynamically generated
Generate monolog config dynamically from $wgDebugLogGroups rather than
having a static point in time snapshot of it. As long we are going to
stick with MWLoggerMonologHandler and MWLoggerMonologLegacyFormatter for
backwards compatible udp2log output this makes more sense and will
reduce drift/confusion of needing to configure output in two places.

Change-Id: I76d9953463d4372c5b35cfa97b413f715ca9d551