Configure logging to use MWLoggerMonologSpi
Change the production logging configuration from the default
MWLoggerLegacySpi to MWLoggerMonologSpi. Configure Monolog to emit log
events as udp2log compatible packets as well as forwarding to Logstash
via a Redis queue. The Monolog configuration is built dynamically based
on the existing $wgDebugLogGroups configuration used for
MWLoggerLegacySpi to avoid confusion/omission that would be likely with
2 sets of logging configuration to manage.

Bug: T76759
Change-Id: I256f99591284020ee6b17d896d505292e61b8866