Optional MWLoggerMonologSpi configuration
Build MWLoggerMonologSpi configuration from $wgDebugLogGroups that will
emit log events as udp2log compatible packets as well as forwarding to
Logstash via a Redis queue. The configuration is built dynamically based
on the existing configuration needed for MWLoggerLegacySpi to avoid
confusion/omission that would be likely with 2 sets of logging
configuration to manage. The $wgDebugLogGroups is still needed to
configure the MWLoggerMonologHandler classes used in the
MWLoggerMonologSpi configuration.

Use is gated by the $wmgUseMonologLogger feature flag in
InitialiseSettings.php that can be used to selectively enable
MWLoggerMonologSpi based logging on a wiki by wiki basis.

Bug: T76759
Change-Id: I720f2cb5975aac40e7c1186c942d9126d67bd190