Remove unused mobile CentralNotice URLs
In WMF setup, it is not expected that EnterMobileMode will be called
from a load.php context. The code removed here apparently switched
in mobile URLs in CentralNotice global variables that contain
CentralNotice URLs for various banner-related purposes. However,
those variables are used almost exclusively in JS, sent in via the
ResourceLoaderGetConfigVars hook.

The only exception is $wgCentralHost, which is used to set a
dns-prefetch in the head. However, since we'll now use instead
non-mobile URLs that don't redirect (see
https://gerrit.wikimedia.org/r/#/c/182078/) it is preferable
to have the non-mobile URL in that link.

That the removed code is non-functional can be seen in the
redirects in background calls to meta on production.

Also removed a comment with a link to a bug that is now
resolved.

Change-Id: I6b74c4a1c7fd1cd03360b0d8a4d97b457997bfdc