Update patch set 1
Patch Set 1: -Code-Review

Has what been communicated to whom?

The stewards are aware of these tools since their launch on beta labs and some chose to participate in the beta testing. News about these tools being tested also went out in a couple different Engineering monthly updates.

The main function of these features is to assist with an anticipated heavy load of rename requests once accounts that will be affected by SUL finalization are contacted. After a few months of testing in labs, the features need to be moved into testing *before* the anticipated heavy load hits so all the stewards and global renamers can get used to the functions available in the rename request queue, as well as getting users to try the request form for processing as well.

Patch-set: 1