conf: Clean up highlight.php and fix http-redirect bug
Follows-up 26929a2f2e4.

https://noc.wikimedia.org/conf/highlight.php redirects to HTTP
due to a missing 'else' statement in the $protocol logic. So the
second block always executes due to the second block just being
an unnamed block that unconditionally executes.

Since we hardcode HTTP 1.1 and because this is primarily a
developer endpoint, just use relative urls for the redirect.

This way it also works locally for developers that have
mediawiki-config itself mounted on a local webserver as it now
is relative to the path instead of hardcoding /conf.

It previously allowed for the hostname to be overridden (why?)
via SERVER_NAME, but did not the path (via e.g. SCRIPT_FILENAME).

Change-Id: Ie15eec3a90fa0a29b965387ee89e74131117fb71