Route all counter stats to tungsten:3811
While HHVM and Zend were running in tandem, we diverted metrics from the HHVM
app servers to a different port on tungsten so we could compare the performance
of HHVM and Zend. Now that the web app servers are uniformly running HHVM,
there is no need for the split, and all stats should go to port 3812.

Change-Id: I50931db3729012c4bece042458b2674e3cc0dafc