Update patch set 3
Patch Set 3: Code-Review+1

So the $wmg* variables are created according to the default values in the 'wmg*' dictionary entries for the current site? This is hard to see in the diff, but assuming this is the case the diff should be ok (I guess otherwise it would fail testing).

Patch-set: 3
Label: Code-Review=+1