Update patch set 1
Patch Set 1:

> Think this is in wrong place.

Depends on: If it should  be disabled on all project except en, test and test2 wiki, then this is correct (but you need to add the configured variable in InitialiseSettings in mobile.php, too, so it will have no effect).

But i think we want to enable it on all projects?

Patch-set: 1
Reviewer: Florianschmidtwelzow <890@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0