Checking "autoreview" instead of "autoconfirmed" for enwiki FlaggedRevs restriction level
Currently, on enwiki "autoconfirmed" is checked for permission to autoreview pages placed under
pending changes at level 1. It causes usergroups with "autoreview" but not "autoconfirmed" to
not being autoreviewed on those. This restores the intended behavior by checking for "autoreview"
instead, which was already granted to all usergroups with "autoconfirmed".

Change-Id: I2ec98e663a75b4d68f2cd2d5889391383b79c478