Update patch set 1
Patch Set 1:

I think the test code should be executing `composer update` rather than `composer install`. If a composer.lock file is present the install command will strictly ensure that the local vendor directory matches the contents of the lock file. In the case of this repo, the versioned vendor directory and associated composer.lock are intended for runtime use rather than testing. This is quite likely to be the case for any repo that we end up having which includes a composer.lock file.

Patch-set: 1