Remove 'autoreview' from 'autoconfirmed', check the former for PC2
Since pending changes protection at level 1 checks for 'autoconfirmed', 'autoreview' had no
effect before commit 106737 where it allowed being autoreviewed when reverting to a stable
version. This was requested for reviewers, but this setting also allows it for any autoconfirmed
user (since all autoconfirmed users are granted 'autoreview'). Yet, when pending changes was
enabled, this behavior was not expected to occur for all autoconfirmed users, since this would
allow those priviledged users to revert to their preferred version (assuming it is reviewed),
giving them an unfair advantage in content disputes. Reviewers on the other hand are trusted
not to do so, so are rollbackers. To address this, 'autoreview' is removed from autoconfirmed
users and 'autoreviewrestore' is granted to rollbackers. In addition, PC2 now checks for
'autoreview' instead of 'review', which makes bots autoreviewed on pages with PC2 and gives a
potential use case to the 'autoreview' usergroup. A commit for FlaggedRevs to make the Sha1
check also for 'autoreviewrestore' is now available at 189668.

Change-Id: I4a2bb4da6aadcaada51734590ca9aa37d11d3cb7