Mobile: Fix JS-Exception for non-english beta projects
Configs should always have a default value and other projects should benefit
from activated Wikidata features, too (otherwise it should be disabled by a feature flag
instead of missing configuration).

Without a default value, the value is null, which results in a JS Exception when
the code tries to access the config as an array.

Bug: T89095
Bug: T88093
Change-Id: I36b62c084b17394ba8942a37f44d84dd26bf6faf