Update patch set 1
Patch Set 1:

We can't fake entries in the compsoer.lock file. It's the list of exactly what composer has put into vendor. So either we pack all the dev requirements around in the repo (and push them to production) or we use update when testing to get the dev requirements on the Jenkins slave.

My recommendation for composer.json in projects is that they use exact version numbers (eg wikimedia/cdb": "1.0.1") to avoid any ambiguity in the version being used. This is a bit in contradiction to the upstream advice from Composer who recommend allowing non-breaking changes (eg wikimedia/cdb": "1.0.*" or "1.*") and the use of the lock file for "repeatable deploys".

Patch-set: 1
Label: Code-Review=0