Update patch set 2
Patch Set 2:

The reason for the revert? As mentioned in the commit message, application/x-gzip wasn't blacklisted to begin with. Thus the original patch removing it from the blacklist was actually not doing anything, and I decided it would be less confusing all around to revert it instead of leaving it in there to continue to not do anything.

The actual problem in T88242 is that a compressed dia file is recognized as "application/x-gzip" rather than "application/x-dia-diagram". Which is technically correct, but not incredibly helpful in this situation.

Patch-set: 2