Update patch set 2
Patch Set 2:

I do not agree, it is very easy to break things with the translate extension, even unintentionally and when doing it in good faith, and fixing such errors is a huge mess and combined with a lot of work. Thus, in my opinion, an opt-in as an additional step is the right thing to implement here, as then actions are taken with even more care and accidental errors are less likely to happen.

Patch-set: 2