Update patch set 3
Patch Set 3:

I've checked FlaggedRevs code and the permission hook checks for the restriction on top of the 'autoreview' userright, and not as I thought as an alternative. So you're correct on this point but that still leaves the issue I'm concerned about, i.e. that any autoconfirmed user can reject changes. I'll make a phabricator task to centralize since it involves two commits. I'll make the FlaggedRevs commit shortly to fix this, then update this one.

Patch-set: 3