Update patch set 1
Patch Set 1:

Can't really comment on this because I don't know the scope of users that would be affected by the change. It seems trivial, but if it breaks notification for some users that would not be good.

Patch-set: 1
Reviewer: Jsahleen <1842@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0