enable MobileFrontend on wikitech
Wikitech has an 'm' DNS name now (I46a4947645f832c5a).

So it seems we also want to enable MobileFrontend.

Also, wikitech was the _only_ exception left here, while
the default is to enable it. Remove that exception.

(This might also make me pick up the idea again to autogenerate the .m. entries
for all wikis in wikimedia.org)

Change-Id: I6b56b3cacda0b17b4d5989ae04c57b66e366a7b2