Set UDP section profiler to write to port 3811
The stats generated by includes/StatsCounter.php are ready for consumption
directly by StatsD, but those generated by includes/profiler/* are not, since
there is currently no way to specify that the unit for timer values should be
milliseconds (as StatsD expects) rather than seconds. So for now the profiler
should be targetting 3811.

There isn't anything listening on that port on graphite1001 at the moment,
but having the profiler write to that port means it stops writing data to StatsD
that StatsD can't parse, and gives us the chance to decide what to do with it
(provision mwprof or, ideally, make the profiler truly StatsD-compatible).

Change-Id: If1c862f95126f27c54e79e61e7eced5d41006cd6