Update patch set 1
Patch Set 1:

> There is needless duplication of handlers.
done.

> But more fundamentally I also have to say I don't like the approach.
The "whole Monolog thing" was actually done so that we could have much greater flexibility in our logging output methods. There was more than one person who objected to the idea of all logs becoming json blobs so just implementing structured json output to go to both fluorine and logstash was not an option. With the Monolog stack we can pick and choose a variety of formatters and transports to get log events out of MediaWiki.

MediaWiki has switched to PSR-3 logging, but not necessarily Monolog logging. I think Monolog is the right thing for the WMF, but it as you have pointed out verbose to configure and potentially complicated. Once Monolog has been proven to work well for our needs I would like to stop using $wgDebugLogGroups and instead just populate $wmgMonologConfig directly. While it is still in a trial stage it seems premature to trigger a "who moved my cheese!" debate about the configuration change so I came up with this method of translating the more familiar $wgDebugLogGroups configuration to the $wmgMonologConfig that is then injected into $wgMWLoggerDefaultSpi.

Eventually I expect the complexity of the Monolog config to be reduced dramatically as we full migrate MediaWiki from using the wfDebugLog calls to direct use of the PSR-3 logging interface. That will bring with it the explicit concept of each log message having a severity level (debug, info, warning, error). When we have that we should be able to configure a default logger that passes messages of warning and error for all channels. Then we would only need to add explicit per-channel configuration when we want info or debug messages or to use sampling. Today however with most messages being sent fir wfDebugLog at info level that would be not very useful.

Patch-set: 1