Update patch set 2
Patch Set 2:

In Iran and China, we don't force users to use https, because a high percentage of users are blocked from using https.

Are users likely to be logged in when they click these links? If so, their forceHTTPS cookie will redirect them to https, and they'll be logged in when they hit the final page.

Patch-set: 2
Reviewer: CSteipp <204@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0