Update patch set 5
Patch Set 5: Code-Review-1

Needs manual rebase.

I also note this conflicts with both of your other patches you have up for SWAT this morning (I51ebbf92 and I8a0fc10f). The best thing to do about that is probably to rebase both of them on top of this one, then this would be merged first; the other alternatives are to have the other two merged first and then manually rebase this one on top of the resulting master during the SWAT window, or to squash all three into one combined patch.

Patch-set: 5
Label: Code-Review=-1