Move sourceswiki special.dblist->wikisource.dblist
sourceswiki is a wikisource: the "default" language-variant
wikisource that predates the language-specific ones and still
in active use for multilingual documents.

Includes handling of the special case in
DbListTests::testDatabaseNamesUseProjectNameAsSuffix
because the database does not match the standard pattern.

InitialiseSettings.php has been modified to remove much of
the special casing of sourceswiki and fold it back within
'wikisource'; but because of the separation some
existing divergent config will be changed:

* Visual Editor will be turned off by default.
  Already the case on every other Wikisource, cleared
  by James F as having been intended.
* DynamicPageList extension will be enabled
  T14423 intended to turn it on for 'all wikisource'
* wmgMFRemovePageActions set to 'upload' disabling
  file uploads.  https://gerrit.wikimedia.org/r/142155
  intended to do so for all wikisources.

At least one divergence remains explicit:

* wgEnotifMinorEdits remains 'true'. (No changeset to
  be found, no bug located).

Bug: T14423
Bug: T68958
Bug: T91534
Change-Id: Idb72066aa4d877753f75e79f6bf0bde4a16e36b1