Enable edit updates for RESTBase on testwiki
RESTBase needs to be updated whenever a new revision is created, or an
old one modified. To that end, we have the RestBaseUpdateJobs extension,
which has been tested in a local setting. With the advent of RESTBase's
public release, it is time to test it on the testwiki to make sure it
performs well.

Bug: T87520
Change-Id: Ie16d1c8718c614031808242c220bd5c793cc18c1