Update patch set 2
Patch Set 2: Code-Review-1

Technically the patch does what it's supposed to do, but from a philosophical point of view, this is in direct and blatant contradiction of what Wikipedia is supposed to stand for.

It has been mentioned multiple times in the past in relation to similar requests that disallowing new page creations violates fundamental principles behind Wikipedia's very existence.

I find it extremely concerning that this is being proposed again after the first request was denied in 2013, and even more concerning that it's being implemented with the reasoning given in Phabricator (ie. OTHERSTUFFEXISTS).

The last time a request like this got fulfilled was Before The Git Era (so 2011), and I think we'll be taking a very slippery slope by starting to accep such requests now.

Therefore this is a -1 from me.

Patch-set: 2