Update patch set 2
Patch Set 2:

@Ori: Well, yeah, I guess you're right.  

I posted that comment just to provide some kind of argument behind the -1. There's a load of comments on the Phabricator task (going back to December 2012), so what I'm saying isn't new.

Anyways, I just wanted to make it known that this is very much opposed, I'll keep discussing the philosophical aspects of this on Phabricator from now on.

Patch-set: 2