Update patch set 1
Patch Set 1:

What all needs migrating? How much duplication is out there? How much needs added to meta? 

I am not sure how much time or announcement this needs. It seems that if we replicate the blacklists as well as possible on meta the transition should be seamless. Am I oversimplifying this?

Patch-set: 1