Use RESTBase for VisualEditor on all wikipedias
Things continue to look good on all wikis deployed so far. This includes
frwiki, which is the wiki with the highest VE usage. In light of this, I think
it makes sense to switch the remaining wikipedias in one go, so that we can
continue monitoring for the remainder of the day.

Change-Id: If7f3e906f56927443df29f0c1cd6a0ba3d36e259