Drop support for 75 languages in SyntaxHighlighter_GeSHi
GeSHi supports 215 languages. The top 6 languages account for more than 50% of
usage, and the top 20 languages for more than 80%. The bottom 75 are not used
at all (or rather: usage of these language's modules was not seen in two hours'
worth of request logs.) Since each supported language gets an entry in
ResourceLoader's start-up module, it makes sense to be economical and drop
support for those languages. (T93025)

Change-Id: Idf34911407bdcd54b904e7d8f11e540ef257c1cc