Tweaks to testwiki/enwiki flaggedrevs config
This adds 'patrol' rights to the reviewer usergroup for testwiki and
enwiki as per T93798. It's for testwiki's sake, but it doesn't hurt to
also give it to enwiki reviewers since autoconfirmed users have 'patrol'
on enwiki.
This also removes the 'autoreview' usergroup which is inoperative since
'autoconfirmed' is also required for autoreview, see the request at
T91934.

Change-Id: I0f842a927d00e13dd674f3b6317b02f39f693732