Update patch set 1
Patch Set 1: Code-Review-1

Hi Odder, Steinsplitter: Just been going through config change requests, found this one. I'm not clear on what our policy for adding domains to this variable is - it's just the licensing of the files it contains, or the trust of the people who will be downloading from it, that we have to be sure of, right?

Also, this change needs to be rebased - it can't merge right now. (-1)

Patch-set: 1
Label: Code-Review=-1