Update patch set 1
Patch Set 1:

Sadly, yes, as far as I know we do :-/ if the extension was loaded on meta it should appear but it isn't (and I don't think there is a plan for it in the near future). We have a couple of these floating through CommonSettings for the same reason (especially regarding flagged revs).

Patch-set: 1
Reviewer: Jalexander <41@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0