Update patch set 1
Patch Set 1:

I don't think that would work. Wikitech does not have a Schema namespace. And the extensions installed on wikitech are not referring to isolated schemas on wikitech. They're referring to the ones from metawiki.

That API url is only used for reading not writing, it should be using the one specified in the 'else' case.

If pointing it to wikitech-api, it will likely result in every event throwing an exception saying the expected schema does not exist in the given API.

Patch-set: 1
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0