Update patch set 5
Patch Set 5: -Code-Review

To recap: the value of the sweeping changes to the logging infrastructure in core is (to my mind) in making MediaWiki behave like a well-mannered modern PHP web application rather than an uncontacted civilization deep in the Amazon rainforests. (After all, the bits that end up flowing over the wire remain mostly unchanged, right?) It is from this point of view that I reviewed this patch, and I gave it a -2 because it seemed to me that the desire to get to a finish line was undermining this value. What good is straight timber if the building you build with it is crooked?

This was my opinion before and it still is my opinion now. I am, however, removing my -2, since it is becoming apparent that we are going in circles. You are free to take or leave my feedback as you like.

Patch-set: 5
Label: Code-Review=0