Update patch set 1
Patch Set 1:

Yuvi: if you wanna write it in python, go for it. I'm still more comfortable/productive in PHP despite it's ugliness.

Guiseppe: Does quilt rely on storing the patches in the repo? Because we have a fresh branch each week they get blown away and I need a completely separate storage that isn't at all tied to the repository the patches apply to.

Honestly this tool is not even required, we can just manage the files manually and it'd be good enough.  This was more of a convenience thing, to make it easier to track them and helpfully, to automatically extract the phab task number from the patch's commit messages.

Patch-set: 1
Reviewer: 20after4 <1706@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0