Update patch set 1
Patch Set 1:

I take it back - it looks like ply would be a good fit for what I need to do: http://pythonhackers.com/p/gochist/ply

quilt could work but it's way more difficult to use and seems like overkill for what I'm trying to achieve.

Really we should be managing these changes as commits in git but that's difficult to manage as well due to the need for them to be kept out of gerrit.

Patch-set: 1