Added *.adlibhosting.com to wgCopyUploadsDomains
This domain is managed by Axiell ALM and offers hosting solutions
to libraries and museums to host collections.

We already allow the Amsterdam Museum collection, and a new request
offered to add the York Museums Trust one. As they target GLAM
institutions, it makes sense to allow the entire domain.

Bug: T95418
Change-Id: Ib8c3f3de2ca181c522891504d1465e2ef748b81d