Update patch set 2
Abandoned

Ok, abandoned: This will likely affect users from Iran and China, which want to translate messages and thus click on notification links. So don't create a extra (but easy) hurdle for them.

If anyone think that HTTPS on first request for all users outweighs that or the rationale is wrong, feel free to restore this change/re-propose this.

Patch-set: 2
Status: abandoned