Update patch set 1
Patch Set 1:

This is my first time to deploy a new extension to wmf and I'm very not sure that this is the right way. So please look carefully and please fix it if there's a any problem. Thank you.

Patch-set: 1
Reviewer: devunt <206@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0