Update patch set 1
Patch Set 1:

So your change does two things:

(i) it cleans our configuration
(ii) it introduces a functional change

That means two commits are required, for each of these two steps. We frown upon . Yes sometimes we add a comma or a space, but only if this is in the SAME block of settings you edit (e.g. if you reorder two lines in this sv.wikivoyage block and not the value before yours).

I would then recommend to create a new change with only the alphabetical order fix (start from a branch against a clean master !), to send it to Gerrit.

Then you can redownload from Gerrit this change and rebase this change against the new one:

git review -d 204473
git rebase <the branch of the new change>
git review

That will create a deps in Gerrit between the two and avoid a merge conflict.
______________

About Gerrit use and +1: 

The goal of the code review is to let at least one another developer read your change.

In other repositories than operations/, you can't +1 a change at all.

In operations/ repositories, you can +1 a change if (i) nobody can review it (ii) you followed a specific process to verify it and you want to indicate that fact (the +1 is then a 'ready to merge and deploy'.

So for configuration changes, as there are active reviewers, don't use +1.

And never, ever +1 your *own* change when the change is rated -1, let the others developers review again instead.

Patch-set: 1