Update patch set 2
Patch Set 2:

As indicated on Phabricator, we would like a confirmation from the community itself.

Responsibilities are clearly shared: the community requests and clarifies the requests, we deploy the clear requests.

When you interpret too much a discussion, you take an active role in the community process and stumble upon their prerogatives. This is not organic.

Furthermore, this change isn't a vital emergency, so there is no drawback to wait upon such confirmation.

Patch-set: 2