Update patch set 1
Patch Set 1:

This is a strange question, this "I don't know where you come from" especially during a review.

I've contributed to this repository since 2012, where I committed 192 changes. As shown in http://git.wikimedia.org/metrics/operations%2Fmediawiki-config.git that means I'm the volunteer with the most commits in this repository.

This is a strange question, because the review process is open to all, and focused to improve quality of our code. We focus on content, not people. We want to improve the quality of our code and ourselves in this process.

This review is not about you, is about how we our configuration file to be cleanly amended.

So my message were simply based on experience and best practices.

And yes, we would really like to see different commits for functional changes and whitespaces/alphabetical order/cleaning issues.

So just tell me if you want to do it or not. If not, and notwithstanding the fact I spent some minutes in the previous review to explain to you how to use git-review and git to properly inserts a commit before another one, I will gladly take care of it.

Patch-set: 1