Update patch set 2
Patch Set 2:

If you don't interpret help as help but instead as commands, and if you reply to recommendations with a "Nobody should tell me what to do!" attitude, I see way bigger problems when it comes to your working together with others in this community. And that was my last comment here.

Patch-set: 2