Update patch set 3
Patch Set 3:

I'm not sure *dropbox is a good idea, since that is quite broad, or at least prefer not to be the person that approves this.

If we can point a wmde subdomain to this dropbox, then that could be option, or transfer the files to some place that wmde or such hosts where the domain is not as broad.

I am not sure how upload via url would work for this.

If we did enable dropbox for this, then maybe only temporary.

Patch-set: 3
Reviewer: Aude <33@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0