Update patch set 2
Patch Set 2:

Reindex is not required for turning it off. I'm not 100% sure if you turn it back on though. You might have to do a reindex then. I'd have to check.

As to what to do with the feature in general I've created T96535 to talk about it. I really don't know what to do with it. Make a check box that turns it on? Make it on by default? Off? I suspect there are lots of folks with opinions and I figure its worth giving them a place to put them.

I don't see a reason to stop this patch though. Nor would I object to opting foundationwiki out because its simple and I don't want to make perfection the enemy of the good. Its wikis like enwiki which I think deserve lots of talking because I'm pretty sure there are features and workflows built on the current behavior. And workflows that the current behavior makes difficult.

Patch-set: 2
Reviewer: Manybubbles <873@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0