Update patch set 1
Patch Set 1:

Let's put this on beta labs https://gerrit.wikimedia.org/r/205474 so Jared can make a call about whether we want to go ahead with this or not.

Personally I'm keen to make this happen and it will help us catch lots of issues with the design.

Unless there are other blocking issues..?

Patch-set: 1