Update patch set 1
Patch Set 1: Code-Review+1

This will enable a beta feature that (when turned on) sets $wgUseMediaWikiUIEverywhere to true?

I'm fine with this (as part of the goal of getting rid of that global).  If above is correct, +2 and sync (even labs config files can not be left merged but undeployed).

Patch-set: 1
Label: Code-Review=+1