monolog: configure wgDebugLogFile for beta cluster
The Monolog conversion attempted to keep the isolation of cli and web
$wgDebugLogFile data streams. Unfortunately the location that the config
for this was placed in (CommonSettings-labs.php) is loaded far too late
in the bootstrap process and ends up not effecting the Monolog
configuration generated by logging.php. Dumb things down by just sending
all wfDebug() messages to a single file (wfDebug.log) for both web and
cli requests.

Also sets the default value of $wgDebugLogFile in prod to /dev/null
which will avoid 'Missing stream uri, the stream can not be opened.'
errors if somehow prod decides to send things to the wgDebugLogFile
Monolog handler.

Bug: T97138
Change-Id: I8651628da95d300ca0602f719f975b4ebd475f72