beta: Import $wmfUdp2logDest global in wmfLabsSettings()
The fix attempted in I264899a doesn't work because the actually beta
cluster only settings array is created inside the wmfLabsSettings()
function rather than the wmfLabsOverrideSettings() function.

I have resisted the temptation to remove that second function entirely,
but I have indented it properly so that it is easier to see what's going
on.

Bug: T95789
Change-Id: I145a3c38a990bcfdc5c021bc89ef4c4002c66d56