Update patch set 2
Patch Set 2:

Hi Alex, I'm not sure why you have -1ed. This is purposely only for English Wikivoyage as we've not rolled out to non-English projects yet (note we are currently discussing/ and planning to with Hebrew and Arabic).

Please read bug report and the bugs / links it includes. We've discussed this with the English Wikivoyage community and they've agreed to trial it.

Patch-set: 2
Label: Verified=0